Clean up and more functionality
