+++ rust/origen/src/generator/processors/mod.rs
@@ -1,0 +1,3 @@
+mod to_string;
+
+pub use to_string::ToString;
