+++ rust/origen/src/lib.rs
@@ -4,9 +4,9 @@ extern crate lazy_static;
 extern crate serde;
 extern crate meta;
 
-mod ast;
 pub mod core;
 pub mod error;
+pub mod generator;
 pub use error::Error;
 
 use self::core::application::config::Config as AppConfig;
